refactor(daemon): macOS-only — drop process.ts + non-macOS service branches
Daemon scope simplified to "macOS LaunchAgent install only". For other
platforms, users run \`isotopes\` in the foreground or supervise it
themselves (tmux, custom systemd unit, etc.).

Removed:
- src/daemon/process.ts + test (the spawn-detached + PID file lifecycle).
  Manual \`isotopes start/stop/status/restart\` commands gone with it.
- src/daemon/service.ts (cross-platform ServiceManager). Replaced by
  src/daemon/launchd.ts — macOS-only, function-based, no class.
- Linux systemd branches (~50 LOC + tests).
- Windows schtasks branches (~100 LOC + tests + Startup folder fallback +
  shell-injection guards that were only needed for the Windows .cmd wrapper).
- ServicePlatform type, getPlatform() — collapsed because only one platform
  remains.
- Now-unused cli.ts helpers: handleDaemonCommand, makeDaemon, formatUptime,
  apiCallSafe.

Added/changed:
- src/daemon/launchd.ts — 6 functions: install, uninstall, enable, disable,
  isInstalled, plus internal helpers. ~115 LOC total.
- cli.ts \`isotopes service\` subcommand now calls launchd directly and
  errors with a clear message on non-darwin hosts.
- CLAUDE.md updated to reflect the macOS-only scope.

Net diff: ~-1000 LOC. Tests 783 → 749 (-34, mostly the deleted
process/service tests; remaining service surface still covered).
