--- src/daemon/process.ts
@@ -1,233 +1,0 @@
-import { spawn, execSync } from "node:child_process";
-import fs from "node:fs/promises";
-import os from "node:os";
-import path from "node:path";
-import { createLogger } from "../logging/logger.js";
-
-const log = createLogger("daemon:process");
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-/** Runtime status snapshot of the daemon process. */
-export interface DaemonStatus {
-  running: boolean;
-  pid?: number;
-  uptime?: number; // seconds
-  startedAt?: Date;
-  configPath?: string;
-}
-
-/** Options for starting the daemon. */
-export interface DaemonOptions {
-  configPath: string;
-  logDir: string;
-  pidFile: string;
-  /** Absolute path to the CLI entry script the daemon should run. */
-  cliEntry: string;
-}
-
-// ---------------------------------------------------------------------------
-// PID file helpers
-// ---------------------------------------------------------------------------
-
-interface PidFileData {
-  pid: number;
-  startedAt?: string;
-}
-
-async function readPidFile(pidFile: string): Promise<PidFileData | undefined> {
-  try {
-    const raw = await fs.readFile(pidFile, "utf-8");
-    const data = JSON.parse(raw.trim()) as PidFileData;
-    return Number.isFinite(data.pid) && data.pid > 0 ? data : undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-async function writePidFile(pidFile: string, data: PidFileData): Promise<void> {
-  await fs.mkdir(path.dirname(pidFile), { recursive: true });
-  await fs.writeFile(pidFile, JSON.stringify(data), "utf-8");
-}
-
-/** Remove pidfile (best-effort). */
-async function removePidFile(pidFile: string): Promise<void> {
-  try {
-    await fs.unlink(pidFile);
-  } catch {
-    // ignore – file may already be gone
-  }
-}
-
-function killProcess(pid: number, force = false): void {
-  if (process.platform === "win32") {
-    try {
-      const flags = force ? "/F /T" : "/T";
-      execSync(`taskkill ${flags} /PID ${pid}`, { stdio: "ignore" });
-    } catch {
-      // Exit code 128 = process not found (already exited) — acceptable.
-    }
-  } else {
-    process.kill(pid, force ? "SIGKILL" : "SIGTERM");
-  }
-}
-
-function isProcessAlive(pid: number): boolean {
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-// ---------------------------------------------------------------------------
-// DaemonProcess
-// ---------------------------------------------------------------------------
-
-/**
- * DaemonProcess — manages the Isotopes daemon as a detached child process.
- *
- * Handles starting (with PID file tracking), graceful stopping (SIGTERM
- * then SIGKILL), status queries, and restart. Log output is redirected
- * to files in the configured log directory.
- */
-export class DaemonProcess {
-  constructor(private options: DaemonOptions) {}
-
-  /**
-   * Start the daemon as a detached child process.
-   * Resolves with the child PID.
-   * Throws if a daemon is already running.
-   */
-  async start(): Promise<{ pid: number }> {
-    if (await this.isRunning()) {
-      const data = await readPidFile(this.options.pidFile);
-      throw new Error(`Daemon already running (pid ${data?.pid})`);
-    }
-
-    // Ensure log directory exists
-    await fs.mkdir(this.options.logDir, { recursive: true });
-
-    const outLog = path.join(this.options.logDir, "isotopes.out.log");
-    const errLog = path.join(this.options.logDir, "isotopes.err.log");
-
-    const outFd = await fs.open(outLog, "a");
-    const errFd = await fs.open(errLog, "a");
-
-    const child = spawn(process.argv[0], [this.options.cliEntry], {
-      detached: true,
-      stdio: ["ignore", outFd.fd, errFd.fd],
-      // Set cwd to user home to avoid inheriting caller's cwd.
-      // Without this, file tools resolve relative paths against the
-      // caller's cwd, which may be another agent's workspace (#92).
-      cwd: os.homedir(),
-      env: {
-        ...process.env,
-        ISOTOPES_DAEMON: "1",
-      },
-    });
-
-    const pid = child.pid;
-    if (pid === undefined) {
-      throw new Error("Failed to start daemon: no PID returned");
-    }
-
-    child.unref();
-
-    // Persist PID and start timestamp
-    const now = new Date();
-    await writePidFile(this.options.pidFile, { pid, startedAt: now.toISOString() });
-
-    // Close log file descriptors in *this* process – the child inherited them
-    await outFd.close();
-    await errFd.close();
-
-    log.info(`Daemon started (pid ${pid})`);
-    return { pid };
-  }
-
-  /**
-   * Stop the running daemon.  Sends SIGTERM first, waits up to 5 s, then
-   * escalates to SIGKILL.
-   */
-  async stop(): Promise<void> {
-    const data = await readPidFile(this.options.pidFile);
-    if (data === undefined || !isProcessAlive(data.pid)) {
-      await removePidFile(this.options.pidFile);
-      throw new Error("Daemon is not running");
-    }
-
-    // Graceful stop
-    killProcess(data.pid);
-
-    // Wait for exit (poll every 100 ms, max 5 s)
-    const deadline = Date.now() + 5_000;
-    while (Date.now() < deadline) {
-      if (!isProcessAlive(data.pid)) {
-        await removePidFile(this.options.pidFile);
-        log.info(`Daemon stopped (pid ${data.pid})`);
-        return;
-      }
-      await new Promise((r) => setTimeout(r, 100));
-    }
-
-    // Force kill
-    try {
-      killProcess(data.pid, true);
-    } catch {
-      // already gone
-    }
-    await removePidFile(this.options.pidFile);
-    log.warn(`Daemon force-killed (pid ${data.pid})`);
-  }
-
-  /**
-   * Return the current daemon status.
-   */
-  async status(): Promise<DaemonStatus> {
-    const data = await readPidFile(this.options.pidFile);
-    if (data === undefined || !isProcessAlive(data.pid)) {
-      if (data !== undefined) {
-        await removePidFile(this.options.pidFile);
-      }
-      return { running: false };
-    }
-
-    const startedAt = data.startedAt ? new Date(data.startedAt) : undefined;
-    const uptime =
-      startedAt !== undefined && !isNaN(startedAt.getTime())
-        ? Math.floor((Date.now() - startedAt.getTime()) / 1_000)
-        : undefined;
-
-    return {
-      running: true,
-      pid: data.pid,
-      uptime,
-      startedAt,
-      configPath: this.options.configPath,
-    };
-  }
-
-  /**
-   * Restart the daemon (stop then start).
-   */
-  async restart(): Promise<{ pid: number }> {
-    try {
-      await this.stop();
-    } catch {
-      // daemon may not be running – that's fine
-    }
-    return this.start();
-  }
-
-  /**
-   * Quick check: is the daemon process alive?
-   */
-  async isRunning(): Promise<boolean> {
-    const data = await readPidFile(this.options.pidFile);
-    return data !== undefined && isProcessAlive(data.pid);
-  }
-}
